pyproject.toml
[project]
name = "tributary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "AbhishekMandapmalvi", email = "abhishekrmandapmalvi@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiobotocore>=3.4.0",
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.5",
    "asyncpg>=0.31.0",
    "azure-storage-blob>=12.28.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.6",
    "click>=8.3.2",
    "cohere>=5.21.1",
    "gcloud-aio-storage>=9.6.4",
    "lxml>=6.0.2",
    "markdown-it-py>=4.0.0",
    "mdit-plain>=1.0.1",
    "nltk>=3.9.4",
    "openai>=2.30.0",
    "pgvector>=0.4.2",
    "pinecone>=8.1.1",
    "pymupdf>=1.27.2.2",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.1",
    "structlog>=25.5.0",
    "tiktoken>=0.12.0",
]

[project.scripts]
tributary = "tributary.cli:cli"

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "aiohttp>=3.13.5",
    "certifi>=2026.2.25",
    "moto[s3]>=5.1.22",
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "requests>=2.33.1",
    "types-aiofiles>=25.1.0.20260408",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
